modify whiteboard component height
